requirements.txt
marker-pdf
yq
python-dotenv
google-generativeai
datasets